General Social Study (GSS) 2008

This dataset includes 6 manifest items about abortion and several covariates from 355 respondents to the 2008 General Social Survey. Respondents answer the questions whether or not think it should be possible for a pregnant woman to obtain a legal abortion. The covariates include age, sex, race, region, and degree of respondents.

Format

A data frame with 355 observations on 11 variables.

DEFECT

If there is a strong chance of serious defect in the baby?

HLTH

If the womans own health is seriously endangered by the pregnancy?

RAPE

If she became pregnant as a result of rape?

POOR

If the family has a very low income and cannot afford any more children?

SINGLE

If she is not married and does not want to marry the man?

NOMORE

If she is married and does not want any more children?

AGE

Respondent's age

SEX

Respondent's race

RACE

Respondent's sex

REGION

Region of interview

DEGREE

Respondent's degree

Source

http://gss.norc.org

References

Smith, Tom W, Peter Marsden, Michael Hout, and Jibum Kim. General Social Surveys, 2008/Principal Investigator, Tom W. Smith; Co-Principal Investigator, Peter V. Marsden; Co-Principal Investigator, Michael Hout; Sponsored by National Science Foundation. -NORC ed.- Chicago: NORC at the University of Chicago

Examples

data("gss")

#> Warning: data set ‘gss’ not found

# Model 1: LCA
lca = glca(item(DEFECT, HLTH, RAPE, POOR, SINGLE, NOMORE) ~ 1,
           data = gss08, nclass = 3)

#> Manifest items :
#>  DEFECT HLTH RAPE POOR SINGLE NOMORE 
#> 
#> Deleted observation(s) : 
#> 3 observation(s) for missing all manifest items
#> 0 observation(s) for missing at least 1 covariates
#> 
#> SET : 1  .....loglik : -687.4486 
#> SET : 2  .....loglik : -687.4488 
#> SET : 3  .....loglik : -693.8564 
#> SET : 4  .....loglik : -687.4486 
#> SET : 5  .....loglik : -687.5667 
#> SET : 6  .....loglik : -687.4486 
#> SET : 7  .....loglik : -687.4488 
#> SET : 8  .....loglik : -687.4486 
#> SET : 9  .....loglik : -687.449 
#> SET : 10 .....loglik : -687.4532 
#> 
#> Start with SET 8 (-687.4486)
#> 
#> Standard LCA Fitting...
#> 
#>  63 iteration 
#> 
#> Converged at 63 iteration (loglik :-687.4486)

summary(lca)

#> 
#> Call:
#> glca(formula = item(DEFECT, HLTH, RAPE, POOR, SINGLE, NOMORE) ~ 
#>     1, data = gss08, nclass = 3)
#> 
#> Manifest items :
#>  DEFECT HLTH RAPE POOR SINGLE NOMORE 
#> 
#> Model : Standard LCA 
#> Number of latent classes : 3 
#> 
#> Mean Prevalence for latent classes for each group:
#>     Class 1 Class 2 Class 3
#> ALL 0.46396 0.34467 0.19137
#> 
#> Number of parameters : 20 
#> 
#> log-likelihood : -687.4486 
#>      G-squared : 29.82695 
#>            AIC : 1414.897 
#>            BIC : 1492.17
#> 
#> Response numbering:
#>        Y = 1 Y = 2
#> DEFECT   YES    NO
#> HLTH     YES    NO
#> RAPE     YES    NO
#> POOR     YES    NO
#> SINGLE   YES    NO
#> NOMORE   YES    NO
#> 
#> Estimated model parameters :
#> Gamma :
#>     Class 1 Class 2 Class 3
#> ALL   0.464  0.3447  0.1914
#> 
#> Rho (Y = 1) :
#>         DEFECT   HLTH   RAPE   POOR SINGLE NOMORE
#> Class 1 1.0000 1.0000 1.0000 0.9813 0.9284 0.9657
#> Class 2 0.8275 0.9453 0.7959 0.0638 0.0390 0.1344
#> Class 3 0.0466 0.3684 0.0949 0.0000 0.0000 0.0000
#> Rho (Y = 2) :
#>         DEFECT   HLTH   RAPE   POOR SINGLE NOMORE
#> Class 1 0.0000 0.0000 0.0000 0.0187 0.0716 0.0343
#> Class 2 0.1725 0.0547 0.2041 0.9362 0.9610 0.8656
#> Class 3 0.9534 0.6316 0.9051 1.0000 1.0000 1.0000

# Model 2: LCA with a covariate
lcr = glca(item(DEFECT, HLTH, RAPE, POOR, SINGLE, NOMORE) ~ SEX,
           data = gss08, nclass = 3)

#> Manifest items :
#>  DEFECT HLTH RAPE POOR SINGLE NOMORE 
#> Covariates : 
#>  SEX 
#> 
#> Deleted observation(s) : 
#> 3 observation(s) for missing all manifest items
#> 0 observation(s) for missing at least 1 covariates
#> 
#> SET : 1  .....loglik : -680.7823 
#> SET : 2  .....loglik : -680.7823 
#> SET : 3  .....loglik : -680.7841 
#> SET : 4  .....loglik : -680.7824 
#> SET : 5  .....loglik : -680.7823 
#> SET : 6  .....loglik : -680.7824 
#> SET : 7  .....loglik : -680.7874 
#> SET : 8  .....loglik : -680.7823 
#> SET : 9  .....loglik : -680.7859 
#> SET : 10 .....loglik : -680.7823 
#> 
#> Start with SET 2 (-680.7823)
#> 
#> Standard LCA with Covariates Fitting...
#> 
#>  70 iteration 
#> 
#> Converged at 70 iteration (loglik :-680.7823)

summary(lcr)

#> 
#> Call:
#> glca(formula = item(DEFECT, HLTH, RAPE, POOR, SINGLE, NOMORE) ~ 
#>     SEX, data = gss08, nclass = 3)
#> 
#> Manifest items :
#>  DEFECT HLTH RAPE POOR SINGLE NOMORE 
#> Covariates : 
#>  SEX 
#> 
#> Model : Standard LCA with Covariates 
#> Number of latent classes : 3 
#> 
#> Mean Prevalence for latent classes for each group:
#>     Class 1 Class 2 Class 3
#> ALL 0.19485 0.46421 0.34093
#> 
#> Number of parameters : 22 
#> 
#> log-likelihood : -680.7823 
#>      G-squared : 57.97919 
#>            AIC : 1405.565 
#>            BIC : 1490.564
#> 
#> Response numbering:
#>        Y = 1 Y = 2
#> DEFECT   YES    NO
#> HLTH     YES    NO
#> RAPE     YES    NO
#> POOR     YES    NO
#> SINGLE   YES    NO
#> NOMORE   YES    NO
#> 
#> Estimated model parameters :
#> Beta :
#>             Class 1/3 Class 2/3
#> (Intercept)   -1.2067    0.3958
#> SEXFEMALE      1.0176   -0.1766
#> 
#> Rho (Y = 1) :
#>         DEFECT   HLTH   RAPE   POOR SINGLE NOMORE
#> Class 1 0.0581 0.3756 0.0922 0.0000 0.0000 0.0000
#> Class 2 1.0000 1.0000 1.0000 0.9812 0.9281 0.9656
#> Class 3 0.8286 0.9464 0.8030 0.0640 0.0393 0.1356
#> Rho (Y = 2) :
#>         DEFECT   HLTH   RAPE   POOR SINGLE NOMORE
#> Class 1 0.9419 0.6244 0.9078 1.0000 1.0000 1.0000
#> Class 2 0.0000 0.0000 0.0000 0.0188 0.0719 0.0344
#> Class 3 0.1714 0.0536 0.1970 0.9360 0.9607 0.8644

coef(lcr)

#> Class 1 / 3 :
#>             Odds Ratio Coefficient  Std. Error  t value  Pr(>|t|)    
#> (Intercept)     0.2992     -1.2067      0.3354   -3.597  0.000492 ***
#> SEXFEMALE       2.7666      1.0176      0.3944    2.580  0.011262 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Class 2 / 3 :
#>             Odds Ratio Coefficient  Std. Error  t value  Pr(>|t|)  
#> (Intercept)     1.4855      0.3958      0.1822    2.172    0.0321 *
#> SEXFEMALE       0.8381     -0.1766      0.2611   -0.676    0.5003  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# Model 3: MGLCA
mglca = glca(item(DEFECT, HLTH, RAPE, POOR, SINGLE, NOMORE) ~ 1,
             group = REGION, data = gss08, nclass = 3)

#> Manifest items :
#>  DEFECT HLTH RAPE POOR SINGLE NOMORE 
#> Grouping variable : REGION 
#> 
#> Deleted observation(s) : 
#> 3 observation(s) for missing all manifest items
#> 0 observation(s) for missing at least 1 covariates
#> 
#> SET : 1  .....loglik : -680.8858 
#> SET : 2  .....loglik : -667.0649 
#> SET : 3  .....loglik : -667.0649 
#> SET : 4  .....loglik : -667.0649 
#> SET : 5  .....loglik : -668.2611 
#> SET : 6  .....loglik : -667.0653 
#> SET : 7  .....loglik : -667.0649 
#> SET : 8  .....loglik : -667.0649 
#> SET : 9  .....loglik : -667.065 
#> SET : 10 .....loglik : -667.0649 
#> 
#> Start with SET 10 (-667.0649)
#> 
#> Multigroup LCA Fitting...
#> 
#>  65 iteration 
#> 
#> Converged at 65 iteration (loglik :-667.0649)

# Model 4: MGLCA with covariates
summary(mglca)

#> 
#> Call:
#> glca(formula = item(DEFECT, HLTH, RAPE, POOR, SINGLE, NOMORE) ~ 
#>     1, group = REGION, data = gss08, nclass = 3)
#> 
#> Manifest items :
#>  DEFECT HLTH RAPE POOR SINGLE NOMORE 
#> Grouping variable : REGION 
#> 
#> Model : Multigroup LCA 
#> Number of latent classes : 3 
#> 
#> Mean Prevalence for latent classes for each group:
#>                 Class 1 Class 2 Class 3
#> NEW ENGLAND     0.00000 0.71336 0.28664
#> MIDDLE ATLANTIC 0.15987 0.13706 0.70307
#> E. NOR. CENTRAL 0.10906 0.42665 0.46428
#> W. NOR. CENTRAL 0.21327 0.26556 0.52117
#> SOUTH ATLANTIC  0.30461 0.27324 0.42216
#> E. SOU. CENTRAL 0.35920 0.50597 0.13484
#> W. SOU. CENTRAL 0.25716 0.50605 0.23679
#> MOUNTAIN        0.14423 0.43191 0.42386
#> PACIFIC         0.12720 0.33324 0.53956
#> 
#> Number of parameters : 36 
#> Number of groups : 9 
#> 
#> log-likelihood : -667.0649 
#>      G-squared : 149.4701 
#>            AIC : 1406.13 
#>            BIC : 1545.221
#> 
#> Response numbering:
#>        Y = 1 Y = 2
#> DEFECT   YES    NO
#> HLTH     YES    NO
#> RAPE     YES    NO
#> POOR     YES    NO
#> SINGLE   YES    NO
#> NOMORE   YES    NO
#> 
#> Estimated model parameters :
#> Gamma :
#>                 Class 1 Class 2 Class 3
#> NEW ENGLAND      0.0000  0.7134  0.2866
#> MIDDLE ATLANTIC  0.1599  0.1371  0.7031
#> E. NOR. CENTRAL  0.1091  0.4267  0.4643
#> W. NOR. CENTRAL  0.2133  0.2656  0.5212
#> SOUTH ATLANTIC   0.3046  0.2732  0.4222
#> E. SOU. CENTRAL  0.3592  0.5060  0.1348
#> W. SOU. CENTRAL  0.2572  0.5061  0.2368
#> MOUNTAIN         0.1442  0.4319  0.4239
#> PACIFIC          0.1272  0.3332  0.5396
#> 
#> Rho (Y = 1) :
#>         DEFECT   HLTH   RAPE   POOR SINGLE NOMORE
#> Class 1 0.0513 0.3649 0.0737 0.0000 0.0000 0.0000
#> Class 2 0.8156 0.9441 0.7982 0.0624 0.0384 0.1321
#> Class 3 1.0000 1.0000 1.0000 0.9806 0.9270 0.9656
#> Rho (Y = 2) :
#>         DEFECT   HLTH   RAPE   POOR SINGLE NOMORE
#> Class 1 0.9487 0.6351 0.9263 1.0000 1.0000 1.0000
#> Class 2 0.1844 0.0559 0.2018 0.9376 0.9616 0.8679
#> Class 3 0.0000 0.0000 0.0000 0.0194 0.0730 0.0344

mglcr = glca(item(DEFECT, HLTH, RAPE, POOR, SINGLE, NOMORE) ~ AGE,
             group = SEX, data = gss08, nclass = 3)

#> Error: `levels.haven_labelled()` not supported.

summary(mglcr)

#> Error in summary(mglcr): object 'mglcr' not found

coef(mglcr)

#> Error in coef(mglcr): object 'mglcr' not found