National Youth Tobacco Survey (NYTS) 2018 This dataset includes 5 manifest items about abortion and several covariates. From the original 2018 National Youth Tobacco Survey data, the Non Hispanic, white students are selected and schools with 30-50 students were selected. Thus, the dataset has 1743 respondents. The covariates include the sex of the respondents and the school ID to which the respondnets belong, and the level of the corresponding school.

National Youth Tobacco Survey (NYTS) 2018 This dataset includes 5 manifest items about abortion and several covariates. From the original 2018 National Youth Tobacco Survey data, the Non Hispanic, white students are selected and schools with 30-50 students were selected. Thus, the dataset has 1743 respondents. The covariates include the sex of the respondents and the school ID to which the respondnets belong, and the level of the corresponding school.

Format

A data frame with 1743 observations on the following 8 variables.

ECIGT

Whether to have tried cigarette smoking, even one or two puffs

ECIGAR

Whether to have ever tried cigar smoking, even one or two puffs

ESLT

Whether to have used chewing tobacco, snuff, or dip

EELCIGT

Whether to have used electronic cigarettes or e-cigarettes

EHOOKAH

Whether to have tried smoking tobacco from a hookah or a waterpipe

SEX

Respondent's Sex

SCH_ID

School ID to which the respondent belongs

SCH_LEV

Level of the corresponding school

Source

https://www.cdc.gov/tobacco/data_statistics/surveys/nyts/index.htm

Examples

data("nyts18")
# \donttest{
# Model 1: LCA
lca = glca(item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ 1,
           data = nyts18, nclass = 3)

#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> 
#> Deleted observation(s) : 
#> 0 observation(s) for missing all manifest items
#> 0 observation(s) for missing at least 1 covariates
#> 
#> SET : 1  .....loglik : -2120.019 
#> SET : 2  .....loglik : -2087.427 
#> SET : 3  .....loglik : -2086.929 
#> SET : 4  .....loglik : -2098.611 
#> SET : 5  .....loglik : -2088.576 
#> SET : 6  .....loglik : -2087.655 
#> SET : 7  .....loglik : -2116.543 
#> SET : 8  .....loglik : -2088.649 
#> SET : 9  .....loglik : -2088.469 
#> SET : 10 .....loglik : -2088.709 
#> 
#> Start with SET 3 (-2086.929)
#> 
#> Standard LCA Fitting...
#> 
#> .. 262 iteration 
#> 
#> Converged at 262 iteration (loglik :-2086.857)

summary(lca)

#> 
#> Call:
#> glca(formula = item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ 
#>     1, data = nyts18, nclass = 3)
#> 
#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> 
#> Model : Standard LCA 
#> Number of latent classes : 3 
#> 
#> Mean Prevalence for latent classes for each group:
#>     Class 1 Class 2 Class 3
#> ALL  0.0446 0.82968 0.12572
#> 
#> Number of parameters : 17 
#> 
#> log-likelihood : -2086.857 
#>      G-squared : 29.52011 
#>            AIC : 4207.714 
#>            BIC : 4300.503
#> 
#> Response numbering:
#>         Y = 1 Y = 2
#> ECIGT     Yes    No
#> ECIGAR    Yes    No
#> ESLT      Yes    No
#> EELCIGT   Yes    No
#> EHOOKAH   Yes    No
#> 
#> Estimated model parameters :
#> Gamma :
#>     Class 1 Class 2 Class 3
#> ALL  0.0446  0.8297  0.1257
#> 
#> Rho (Y = 1) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.9340 0.9777 0.6542  0.9767  0.6776
#> Class 2 0.0125 0.0074 0.0102  0.0843  0.0071
#> Class 3 0.5308 0.3655 0.1901  0.8717  0.0618
#> Rho (Y = 2) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.0660 0.0223 0.3458  0.0233  0.3224
#> Class 2 0.9875 0.9926 0.9898  0.9157  0.9929
#> Class 3 0.4692 0.6345 0.8099  0.1283  0.9382

# Model 2: LCR
lca = glca(item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ SEX,
           data = nyts18, nclass = 3)

#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> Covariates : 
#>  SEX 
#> 
#> Deleted observation(s) : 
#> 0 observation(s) for missing all manifest items
#> 0 observation(s) for missing at least 1 covariates
#> 
#> SET : 1  .....loglik : -2085.646 
#> SET : 2  .....loglik : -2087.207 
#> SET : 3  .....loglik : -2085.573 
#> SET : 4  .....loglik : -2112.271 
#> SET : 5  .....loglik : -2085.142 
#> SET : 6  .....loglik : -2090.27 
#> SET : 7  .....loglik : -2089.19 
#> SET : 8  .....loglik : -2083.955 
#> SET : 9  .....loglik : -2085.167 
#> SET : 10 .....loglik : -2084.187 
#> 
#> Start with SET 8 (-2083.955)
#> 
#> Standard LCA with Covariates Fitting...
#> 
#> ... 325 iteration 
#> 
#> Converged at 325 iteration (loglik :-2083.746)

summary(lca)

#> 
#> Call:
#> glca(formula = item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ 
#>     SEX, data = nyts18, nclass = 3)
#> 
#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> Covariates : 
#>  SEX 
#> 
#> Model : Standard LCA with Covariates 
#> Number of latent classes : 3 
#> 
#> Mean Prevalence for latent classes for each group:
#>     Class 1 Class 2 Class 3
#> ALL 0.12821 0.82078 0.05102
#> 
#> Number of parameters : 19 
#> 
#> log-likelihood : -2083.746 
#>      G-squared : 98.82476 
#>            AIC : 4205.492 
#>            BIC : 4309.197
#> 
#> Response numbering:
#>         Y = 1 Y = 2
#> ECIGT     Yes    No
#> ECIGAR    Yes    No
#> ESLT      Yes    No
#> EELCIGT   Yes    No
#> EHOOKAH   Yes    No
#> 
#> Estimated model parameters :
#> Beta :
#>             Class 1/3 Class 2/3
#> (Intercept)    1.3004    3.1542
#> SEXFemale     -0.6691   -0.6633
#> 
#> Rho (Y = 1) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.4939 0.3160 0.1632  0.8466  0.0554
#> Class 2 0.0109 0.0069 0.0099  0.0787  0.0069
#> Class 3 0.9102 0.9695 0.6368  0.9786  0.6113
#> Rho (Y = 2) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.5061 0.6840 0.8368  0.1534  0.9446
#> Class 2 0.9891 0.9931 0.9901  0.9213  0.9931
#> Class 3 0.0898 0.0305 0.3632  0.0214  0.3887

coef(lca)

#> Class 1 / 3 :
#>             Odds Ratio Coefficient  Std. Error  t value  Pr(>|t|)    
#> (Intercept)     3.6707      1.3004      0.3032    4.289  9.68e-05 ***
#> SEXFemale       0.5121     -0.6691      0.3607   -1.855    0.0703 .  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Class 2 / 3 :
#>             Odds Ratio Coefficient  Std. Error  t value  Pr(>|t|)    
#> (Intercept)    23.4332      3.1542      0.2765   11.408  9.86e-15 ***
#> SEXFemale       0.5151     -0.6633      0.2855   -2.323    0.0249 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# Model 3: MGLCA
mglca = glca(item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ 1,
             group = SEX, data = nyts18, nclass = 3)

#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> Grouping variable : SEX 
#> 
#> Deleted observation(s) : 
#> 0 observation(s) for missing all manifest items
#> 0 observation(s) for missing at least 1 covariates
#> 
#> SET : 1  .....loglik : -2084.171 
#> SET : 2  .....loglik : -2113.358 
#> SET : 3  .....loglik : -2108.645 
#> SET : 4  .....loglik : -2113.648 
#> SET : 5  .....loglik : -2113.485 
#> SET : 6  .....loglik : -2113.49 
#> SET : 7  .....loglik : -2112.596 
#> SET : 8  .....loglik : -2113.462 
#> SET : 9  .....loglik : -2084.669 
#> SET : 10 .....loglik : -2084.384 
#> 
#> Start with SET 1 (-2084.171)
#> 
#> Multigroup LCA Fitting...
#> 
#> ... 325 iteration 
#> 
#> Converged at 325 iteration (loglik :-2083.746)

summary(mglca)

#> 
#> Call:
#> glca(formula = item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ 
#>     1, group = SEX, data = nyts18, nclass = 3)
#> 
#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> Grouping variable : SEX 
#> 
#> Model : Multigroup LCA 
#> Number of latent classes : 3 
#> 
#> Mean Prevalence for latent classes for each group:
#>        Class 1 Class 2 Class 3
#> Male   0.83380 0.13061 0.03558
#> Female 0.80738 0.12574 0.06689
#> 
#> Number of parameters : 19 
#> Number of groups : 2 
#> 
#> log-likelihood : -2083.746 
#>      G-squared : 98.82476 
#>            AIC : 4205.492 
#>            BIC : 4309.197
#> 
#> Response numbering:
#>         Y = 1 Y = 2
#> ECIGT     Yes    No
#> ECIGAR    Yes    No
#> ESLT      Yes    No
#> EELCIGT   Yes    No
#> EHOOKAH   Yes    No
#> 
#> Estimated model parameters :
#> Gamma :
#>        Class 1 Class 2 Class 3
#> Male    0.8338  0.1306  0.0356
#> Female  0.8074  0.1257  0.0669
#> 
#> Rho (Y = 1) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.0109 0.0069 0.0099  0.0787  0.0069
#> Class 2 0.4939 0.3160 0.1632  0.8466  0.0554
#> Class 3 0.9102 0.9695 0.6368  0.9786  0.6113
#> Rho (Y = 2) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.9891 0.9931 0.9901  0.9213  0.9931
#> Class 2 0.5061 0.6840 0.8368  0.1534  0.9446
#> Class 3 0.0898 0.0305 0.3632  0.0214  0.3887

# Model 4: MLCA
mlca = glca(item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ 1,
   group = SCH_ID, data = nyts18, nclass = 3, ncluster = 2)

#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> Grouping variable : SCH_ID 
#> 
#> Deleted observation(s) : 
#> 0 observation(s) for missing all manifest items
#> 0 observation(s) for missing at least 1 covariates
#> 
#> SET : 1  .....loglik : -1967.699 
#> SET : 2  .....loglik : -1955.501 
#> SET : 3  .....loglik : -1961.822 
#> SET : 4  .....loglik : -1957.145 
#> SET : 5  .....loglik : -1955.624 
#> SET : 6  .....loglik : -1958.154 
#> SET : 7  .....loglik : -1955.497 
#> SET : 8  .....loglik : -1982.574 
#> SET : 9  .....loglik : -1955.49 
#> SET : 10 .....loglik : -1962.948 
#> 
#> Start with SET 9 (-1955.49)
#> 
#> Multilevel LCA Fitting...
#> 
#> . 128 iteration 
#> 
#> Converged at 128 iteration (loglik :-1955.487)

summary(mlca)

#> 
#> Call:
#> glca(formula = item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ 
#>     1, group = SCH_ID, data = nyts18, nclass = 3, ncluster = 2)
#> 
#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> Grouping variable : SCH_ID 
#> 
#> Model : Multilevel LCA 
#> Number of latent classes : 3 
#> Number of latent clusters : 2 
#> 
#> Mean Prevalence for latent clusters:
#> Cluster 1 Cluster 2 
#>    0.6207    0.3793 
#> 
#> Mean Prevalence for latent classes:
#>           Class 1 Class 2 Class 3
#> Cluster 1 0.06129 0.00876 0.92995
#> Cluster 2 0.34686 0.14137 0.51177
#> 
#> 
#> Number of parameters : 20 
#> Number of groups : 45 
#> 
#> log-likelihood : -1955.487 
#>      G-squared : 768.5035 
#>            AIC : 3950.973 
#>            BIC : 4060.137
#> 
#> Response numbering:
#>         Y = 1 Y = 2
#> ECIGT     Yes    No
#> ECIGAR    Yes    No
#> ESLT      Yes    No
#> EELCIGT   Yes    No
#> EHOOKAH   Yes    No
#> 
#> Estimated model parameters :
#> Delta :
#> Cluster 1 Cluster 2 
#>    0.6207    0.3793 
#> 
#> Gamma :
#>           Class 1 Class 2 Class 3
#> Cluster 1  0.0613  0.0088  0.9300
#> Cluster 2  0.3469  0.1414  0.5118
#> 
#> Rho (Y = 1) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.3489 0.2006 0.1236  0.7783  0.0443
#> Class 2 0.9112 0.9750 0.5651  0.9778  0.5364
#> Class 3 0.0062 0.0043 0.0088  0.0413  0.0057
#> Rho (Y = 2) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.6511 0.7994 0.8764  0.2217  0.9557
#> Class 2 0.0888 0.0250 0.4349  0.0222  0.4636
#> Class 3 0.9938 0.9957 0.9912  0.9587  0.9943
#> 

# Model 5: MLCA with level-1 covariate(s) only
mlcr = glca(item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ SEX,
            group = SCH_ID, data = nyts18, nclass = 3, ncluster = 2)

#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> Grouping variable : SCH_ID 
#> Covariates (Level 1) : 
#>  SEX 
#> 
#> Deleted observation(s) : 
#> 0 observation(s) for missing all manifest items
#> 0 observation(s) for missing at least 1 covariates
#> 
#> SET : 1  .....loglik : -1951.881 
#> SET : 2  .....loglik : -1973.677 
#> SET : 3  .....loglik : -1953.913 
#> SET : 4  .....loglik : -1952.26 
#> SET : 5  .....loglik : -1951.884 
#> SET : 6  .....loglik : -1980.966 
#> SET : 7  .....loglik : -1951.887 
#> SET : 8  .....loglik : -1955.116 
#> SET : 9  .....loglik : -1953.726 
#> SET : 10 .....loglik : -1956.755 
#> 
#> Start with SET 1 (-1951.881)
#> 
#> Multilevel LCA with Covariates Fitting...
#> 
#> . 146 iteration 
#> 
#> Converged at 146 iteration (loglik :-1951.879)

summary(mlcr)

#> 
#> Call:
#> glca(formula = item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ 
#>     SEX, group = SCH_ID, data = nyts18, nclass = 3, ncluster = 2)
#> 
#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> Grouping variable : SCH_ID 
#> Covariates (Level 1) : 
#>  SEX 
#> 
#> Model : Multilevel LCA with Covariates 
#> Number of latent classes : 3 
#> Number of latent clusters : 2 
#> 
#> Mean Prevalence for latent clusters:
#> Cluster 1 Cluster 2 
#>   0.61682   0.38318 
#> 
#> Mean Prevalence for latent classes:
#>           Class 1 Class 2 Class 3
#> Cluster 1 0.06023 0.93054 0.00923
#> Cluster 2 0.34682 0.51220 0.14098
#> 
#> 
#> Number of parameters : 22 
#> Number of groups : 45 
#> 
#> log-likelihood : -1951.879 
#>      G-squared : 1116.424 
#>            AIC : 3947.757 
#>            BIC : 4067.837
#> 
#> Response numbering:
#>         Y = 1 Y = 2
#> ECIGT     Yes    No
#> ECIGAR    Yes    No
#> ESLT      Yes    No
#> EELCIGT   Yes    No
#> EHOOKAH   Yes    No
#> 
#> Estimated model parameters :
#> Delta :
#> Cluster 1 Cluster 2 
#>    0.6168    0.3832 
#> 
#> Beta (level 1) :
#> $Cluster1
#>             Class 1/3 Class 2/3
#> (Intercept)    2.2229    4.9893
#> SEXFemale     -0.5998   -0.6576
#> 
#> $Cluster2
#>             Class 1/3 Class 2/3
#> (Intercept)    1.2317    1.6489
#> SEXFemale     -0.5998   -0.6576
#> 
#> 
#> Rho (Y = 1) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.3502 0.1987 0.1188  0.7726  0.0438
#> Class 2 0.0059 0.0042 0.0087  0.0416  0.0057
#> Class 3 0.9025 0.9712 0.5744  0.9782  0.5341
#> Rho (Y = 2) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.6498 0.8013 0.8812  0.2274  0.9562
#> Class 2 0.9941 0.9958 0.9913  0.9584  0.9943
#> Class 3 0.0975 0.0288 0.4256  0.0218  0.4659
#> 

coef(mlcr)

#> 
#> Level 1 Coefficients :
#> 
#> Class 1 / 3 :
#>           Odds Ratio Coefficient  Std. Error  t value  Pr(>|t|)    
#> SEXFemale     0.5489     -0.5998      0.1819   -3.297  0.000996 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Class 2 / 3 :
#>           Odds Ratio Coefficient  Std. Error  t value  Pr(>|t|)   
#> SEXFemale     0.5181     -0.6576      0.2071   -3.176   0.00152 **
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# Model 6: MLCA with level-1 and level-2 covariate(s)
# (SEX: level-1 covariate, PARTY: level-2 covariate)
mlcr2 = glca(item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ SEX + SCH_LEV,
             group = SCH_ID, data = nyts18, nclass = 3, ncluster = 2)

#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> Grouping variable : SCH_ID 
#> Covariates (Level 2) : 
#>  SCH_LEV 
#> Covariates (Level 1) : 
#>  SEX 
#> 
#> Deleted observation(s) : 
#> 0 observation(s) for missing all manifest items
#> 0 observation(s) for missing at least 1 covariates
#> 
#> SET : 1  .....loglik : -1939.598 
#> SET : 2  .....loglik : -1943.697 
#> SET : 3  .....loglik : -1940.134 
#> SET : 4  .....loglik : -1920.707 
#> SET : 5  .....loglik : -1920.622 
#> SET : 6  .....loglik : -1942.973 
#> SET : 7  .....loglik : -1920.223 
#> SET : 8  .....loglik : -1921.027 
#> SET : 9  .....loglik : -1928.153 
#> SET : 10 .....loglik : -1919.94 
#> 
#> Start with SET 10 (-1919.94)
#> 
#> Multilevel LCA with Covariates Fitting...
#> 
#> . 190 iteration 
#> 
#> Converged at 190 iteration (loglik :-1919.937)

summary(mlcr2)

#> 
#> Call:
#> glca(formula = item(ECIGT, ECIGAR, ESLT, EELCIGT, EHOOKAH) ~ 
#>     SEX + SCH_LEV, group = SCH_ID, data = nyts18, nclass = 3, 
#>     ncluster = 2)
#> 
#> Manifest items :
#>  ECIGT ECIGAR ESLT EELCIGT EHOOKAH 
#> Grouping variable : SCH_ID 
#> Covariates (Level 2) : 
#>  SCH_LEV 
#> Covariates (Level 1) : 
#>  SEX 
#> 
#> Model : Multilevel LCA with Covariates 
#> Number of latent classes : 3 
#> Number of latent clusters : 2 
#> 
#> Mean Prevalence for latent clusters:
#> Cluster 1 Cluster 2 
#>   0.54083   0.45917 
#> 
#> Mean Prevalence for latent classes:
#>           Class 1 Class 2 Class 3
#> Cluster 1  0.0199 0.89032 0.08978
#> Cluster 2  0.1175 0.59206 0.29045
#> 
#> 
#> Number of parameters : 24 
#> Number of groups : 45 
#> 
#> log-likelihood : -1919.937 
#>      G-squared : 1052.541 
#>            AIC : 3887.874 
#>            BIC : 4018.87
#> 
#> Response numbering:
#>         Y = 1 Y = 2
#> ECIGT     Yes    No
#> ECIGAR    Yes    No
#> ESLT      Yes    No
#> EELCIGT   Yes    No
#> EHOOKAH   Yes    No
#> 
#> Estimated model parameters :
#> Delta :
#> Cluster 1 Cluster 2 
#>    0.5408    0.4592 
#> 
#> Beta (level 1) :
#> $Cluster1
#>             Class 1/3 Class 2/3
#> (Intercept)   -1.5691    1.0654
#> SEXFemale      0.5148   -0.1159
#> 
#> $Cluster2
#>             Class 1/3 Class 2/3
#> (Intercept)   -1.0668   -0.1671
#> SEXFemale      0.5148   -0.1159
#> 
#> Beta (level 2) :
#>                    Class 1/3 Class 2/3
#> SCH_LEVHigh School   -0.6134    1.9522
#> 
#> Rho (Y = 1) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.8914 0.9657 0.5507  0.9782  0.5049
#> Class 2 0.0033 0.0034 0.0074  0.0372  0.0056
#> Class 3 0.3227 0.1696 0.1127  0.7266  0.0394
#> Rho (Y = 2) :
#>          ECIGT ECIGAR   ESLT EELCIGT EHOOKAH
#> Class 1 0.1086 0.0343 0.4493  0.0218  0.4951
#> Class 2 0.9967 0.9966 0.9926  0.9628  0.9944
#> Class 3 0.6773 0.8304 0.8873  0.2734  0.9606
#> 

coef(mlcr2)

#> 
#> Level 1 Coefficients :
#> 
#> Class 1 / 3 :
#>           Odds Ratio Coefficient  Std. Error  t value  Pr(>|t|)  
#> SEXFemale     1.6733      0.5148      0.2180    2.362    0.0183 *
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Class 2 / 3 :
#>           Odds Ratio Coefficient  Std. Error  t value  Pr(>|t|)
#> SEXFemale     0.8906     -0.1159      0.1716   -0.675       0.5
#> 
#> 
#> Level 2 Coefficients :
#> 
#> Class 1 / 3 :
#>                    Odds Ratio Coefficient  Std. Error  t value  Pr(>|t|)
#> SCH_LEVHigh School     0.5415     -0.6134      0.6587   -0.931     0.352
#> 
#> Class 2 / 3 :
#>                    Odds Ratio Coefficient  Std. Error  t value  Pr(>|t|)    
#> SCH_LEVHigh School     7.0439      1.9522      0.4185    4.665  3.33e-06 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# }